Create global helpers header file
Organize JSON helpers in a single file
