+++ .gitignore
@@ -4,3 +4,4 @@
 .vscode/launch.json
 .vscode/ipch
 .vscode
+.DS_Store
